refactor: group services by network side
